chatbots/document-chatbot/requirements.txt
acresaiofilesaltairanyioattrsauthautoscrollblinkerboto3botocorecachetoolscertificharset-normalizerci-infoclickcoloramaconfigobjconfigparserdnspythondotenvetelemetryeventletfalconfilelockFlaskfrontendgitdbGitPythongreenletgunicornh11httplib2idnaisodateitsdangerousJinja2jmespathjsonschemajsonschema-specificationslooseversionlxmlMarkupSafemongoenginemysql-connector-pythonnarwhalsnetworkxnibabelnipypenumpypackagingpandaspathlibpdf2imagepillowprotobufprovpsycopg2-binary  # Changed from psycopg2 to psycopg2-binarypuremagicpyarrowpydeckpydotPyJWTpymongoPyMuPDFpynputpyodbcpyparsingpytesseractpython-dateutilpython-docxpython-dotenvpytilspytzpyxnatrazorpayrdflibreferencingrequestsrpds-pys3transferscipysetuptoolssimplejsonsixsmmapsniffiostarlettestreamlitstreamlit-autorefreshstreamlit-javascripttenacitytomltoolstornadotraitstyping_extensionstzdataurllib3uvicornwatchdogWerkzeug